requirements.txt
python
aiogram
sqlalchemy
python-dotenv